third_party/blink/web_tests/platform/generic/fast/dom/HTMLImageElement/image-src-added-srcset-w-descriptors-expected.txt

PASS img.currentSrc.indexOf("blue_rect.jpg") is not -1
PASS img.currentSrc.indexOf("image-set-4x") is not -1
PASS successfullyParsed is true

TEST COMPLETE
